requirements.txt
azure-core==1.32.0
azure-cosmos==4.8.0
azure-identity==1.19.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
google-auth==2.36.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
requests==2.32.3
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
